+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added in import statements.
 * Added in `external` function attribute.
 * Added in exceptions from the name and type resolvers.
+* Added in `Char` type and character literals.
 
 #### v0.2.0.0